+++ TinyCSV/CSVException.cs
@@ -11,6 +11,10 @@ namespace TinyCSV
         public CSVException(string message) : base(message)
         {
         }
+        
+        public CSVException(Exception innerException) : base(innerException.Message, innerException)
+        {
+        }
 
         public CSVException(string message, Exception innerException) : base(message, innerException)
         {